+++ NEWS.md
@@ -8,6 +8,7 @@
 * `gcs_global_bucket` and `gcs_get_global_bucket` to set global bucket name
 * Add options to set environment variables for API keys, default buckets etc.
 * Add print S3 methods for objects, buckets and ACL
+* Add option to use your own write function for R object file uploads in `gcs_upload`
 
 # googleCloudStorageR 0.1.0